chore(vision): extract vision gui header to it's own component
